Update creation & index HTML
Creation had the wrong link connected to the Lunapic button and the homepage didn't have a link to contact me, so I fixed that as well. The grid at the bottom of the homepage is now in order of first occurrence on the Creation page. All edited files have been validated.